+++ app/src/main/java/floor/planner/util/math/Random.java
@@ -5,14 +5,14 @@ public class Random {
      * Default max random value which should be next value down from 1 since
      * Math.random() returns a number in [0, 1)
      */
-    private static float RAND_MAX = (float) Math.nextDown(1.0);
+    private static double RAND_MAX = (double) Math.nextDown(1.0);
 
-    public static float randomFloat() {
-        return (float) Math.random() / (RAND_MAX + 1);
+    public static double randomDouble() {
+        return (double) Math.random() / (RAND_MAX + 1);
     }
 
-    public static float randomFloat(int min, int max) {
+    public static double randomDouble(int min, int max) {
         // returns a random real in [min, max)
-        return (float) (min + ((max - min) * Math.random()));
+        return (double) (min + ((max - min) * Math.random()));
     }
 }
